cargo: Make all procfs usage limited to Linux
This lets us build all Rust components on macOS.
