+++ subprojects/hydra-queue-runner/Cargo.toml
@@ -47,7 +47,6 @@ http-body-util.workspace = true
 hyper                    = { workspace = true, features = [ "full" ] }
 hyper-util.workspace     = true
 jiff.workspace           = true
-procfs.workspace         = true
 prometheus               = { workspace = true, features = [ "process" ] }
 tower                    = { workspace = true, features = [ "util" ] }
 tower-http               = { workspace = true, features = [ "trace" ] }
@@ -76,6 +75,9 @@ store-transfer.workspace                 = true
 [dev-dependencies]
 harmonia-file-nar = { workspace = true, features = [ "test" ] }
 tempfile          = "3.23.0"
+
+[target.'cfg(target_os = "linux")'.dependencies]
+procfs.workspace = true
 
 [target.'cfg(not(target_env = "msvc"))'.dependencies]
 tikv-jemallocator.workspace = true
